#19 refactor: LogButton -> LifePoopButton 이름 변경, title을 생성자로 받도록 수정, highlighted 색상 설정
